requirements.txt
fastapi==0.95.2
pydantic==1.10.13
starlette==0.27.0
uvicorn[standard]
jinja2
sentence-transformers==3.0.1
faiss-cpu
transformers==4.40.0
requests
gcsfs==2023.6.0
